requirements.txt
cyclonedds==0.10.5
kornia==0.8.1
matplotlib==3.10.5
numpy==1.26.4
open3d==0.19.0
opencv_python==4.11.0.86
Pillow==11.3.0
pycuda==2025.1.1
PyYAML==6.0.2
PyYAML==6.0.2
rclpy==3.3.17
rerun-sdk==0.22.1
scikit_learn==1.7.1
scipy==1.15.3
setuptools==80.9.0
tifffile==2025.5.10
timm==1.0.19
tqdm==4.67.1
transformers==4.53.0
typing_extensions==4.14.1
webdataset==0.2.86
open_clip_torch==2.32.0